% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_plots.R
\name{CreatePdfFile}
\alias{CreatePdfFile}
\title{Luo pdf-tiedoston kuvasta}
\usage{
CreatePdfFile(obj, fname, thiswidth, thisheight)
}
\arguments{
\item{obj}{ggplot-kuvio}

\item{fname}{tallennettava tiedostonimi}

\item{thiswidth}{kuvan leveys}

\item{thisheight}{kuvan korkeus}
}
\description{
Luo pdf-tiedoston kuvasta
}
