% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{GetLabelNamesForGgMcmc}
\alias{GetLabelNamesForGgMcmc}
\title{On the basis of a coda object, retrieve human readable names for the variables}
\usage{
GetLabelNamesForGgMcmc(VarName, VarValues, MeanGuide, MeanTreshold = 0.1,
  lang = "fi")
}
\arguments{
\item{VarName}{tutkittava muuttuja}

\item{VarValues}{lista, jossa arvot}

\item{MeanGuide}{vektori keskiarvoista}

\item{MeanTreshold}{minkä alle jäävät arvot väritetään vaaleammalla värillä}

\item{lang}{tulostetaanko suomen- vai englanninkielisiä kuvioita}

\item{MeanGuide}{vektori keskiarvoista}
}
\description{
On the basis of a coda object, retrieve human readable names for the variables
}
