% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_bayesian.R
\docType{methods}
\name{LoadResults,StandardBayesian-method}
\alias{LoadResults,StandardBayesian-method}
\title{Hakee rds-muodossa tallennetun posterioridatan perusteella tietyt usein tarvittavat kuviot ja 
oliot yhdeksi listaksi. Jos listaa ei ole vielä olemassa, tallentaa tämän rds-muodossa.}
\usage{
\S4method{LoadResults}{StandardBayesian}(object, force_reload, lang = "fi")
}
\arguments{
\item{force_reload}{whether or not to force reloading the plots and all the other stuff}

\item{lang}{minkäkielisiä kuvioita tehdään}
}
\description{
Hakee rds-muodossa tallennetun posterioridatan perusteella tietyt usein tarvittavat kuviot ja 
oliot yhdeksi listaksi. Jos listaa ei ole vielä olemassa, tallentaa tämän rds-muodossa.
}
