% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_bayesian.R
\docType{methods}
\name{PrintSvg,StandardBayesian-method}
\alias{PrintSvg,StandardBayesian-method}
\title{Tulostaa html-esityksen kuvioista}
\usage{
\S4method{PrintSvg}{StandardBayesian}(object, ...)
}
\description{
Tulostaa html-esityksen kuvioista
}
