% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_plots.R
\name{ReadData}
\alias{ReadData}
\title{Lukee tilastollista mallia koskevan datan rds-tiedostosta}
\usage{
ReadData(name)
}
\arguments{
\item{name}{mallin nimi}
}
\description{
Lukee tilastollista mallia koskevan datan rds-tiedostosta
}
