% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_bayesian.R
\docType{methods}
\name{ResetAllData,StandardBayesian-method}
\alias{ResetAllData,StandardBayesian-method}
\title{Poista kaikki tallennettu tieto t채st채 mallista, jotta se olisi esimerkiksi helppo ladata uudestaan muutetuilla parametreilla}
\usage{
\S4method{ResetAllData}{StandardBayesian}(object, forcethis = F)
}
\description{
Poista kaikki tallennettu tieto t채st채 mallista, jotta se olisi esimerkiksi helppo ladata uudestaan muutetuilla parametreilla
}
