% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_bayesian.R
\docType{methods}
\name{SaveOriginalData,StandardBayesian-method}
\alias{SaveOriginalData,StandardBayesian-method}
\title{Testaa, onko jo tallennettuna t채m채n mallin pohjana olevasta datasta rds-tiedostoa. Jos ei ole, tallentaa sellaisen.}
\usage{
\S4method{SaveOriginalData}{StandardBayesian}(object)
}
\description{
Testaa, onko jo tallennettuna t채m채n mallin pohjana olevasta datasta rds-tiedostoa. Jos ei ole, tallentaa sellaisen.
}
