% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_plots.R
\name{CreatePdfFile}
\alias{CreatePdfFile}
\title{Luo pdf-tiedoston kuvasta}
\usage{
CreatePdfFile(obj, fname, thiswidth, thisheight, coord_start, coord_end)
}
\arguments{
\item{obj}{ggplot-kuvio}

\item{fname}{tallennettava tiedostonimi}

\item{thiswidth}{kuvan leveys}

\item{thisheight}{kuvan korkeus}

\item{coord_start}{x-akselin alkup채채}

\item{coord_end}{x-akselin loppup채채}
}
\description{
Luo pdf-tiedoston kuvasta
}
