% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saving_plots.R
\name{SavePdf}
\alias{SavePdf}
\title{Tallentaa pdf-tiedostot yhden til. mallin kuvioista.}
\usage{
SavePdf(datalist, modelname, plotname, thiswidth = F, thisheight = F,
  coords = c(-2, 2, -2, 2))
}
\arguments{
\item{datalist}{StandardBayesian-olion datalista}

\item{modelname}{mallin nimi}

\item{plotname}{kuvion nimi tai muuttujan nimi}

\item{thiswidth}{kuvan leveys}

\item{thisheight}{kuvan korkeus}

\item{coords}{x-akselin raja-arvot muodossa c(general.start,general.end,int.start,int.end)}
}
\description{
Tallentaa pdf-tiedostot yhden til. mallin kuvioista.
}
